requirements.txt
coverage==5.0.4
Flask==1.1.2
pylint==2.4.4
pytest-flask==1.0.0
python-dotenv==0.11.0
yapf==0.29.0
aniso8601==8.0.0
atomicwrites==1.4.0
attrs==19.3.0
cbor==1.0.0
click==7.1.2
colorama==0.4.3
Flask-RESTful==0.3.8
geomet==0.1.2
gunicorn==20.0.4
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2020.1
scylla-driver==3.22.2
six==1.15.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
pandas==1.0.5
flask-cors==3.0.9
schedule==1.1.0
openpyxl==3.0.7
xlsxwriter==3.0.2
psycopg2-binary==2.8.5
schedule==1.1.0
pydomo==0.3.0.9
python-dotenv==1.0.0